requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==5.1.0
dict==2020.12.3
Django==4.0.1
django-bootstrap-form==3.4
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
idna==3.3
ipdb==0.13.9
ipython==7.31.0
jedi==0.18.1
logger==1.4
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
phonenumberslite==8.12.40
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.2
python-telegram==0.15.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
typing==3.7.4.3
urllib3==1.26.8
utils==1.0.1
values==2020.12.3
wcwidth==0.2.5
